feat (openai): openai 의존성 분리
